split text and image generation even further
