+++ bot-config.json
@@ -56,9 +56,7 @@
     "DISCORD": {
         "MAX_CHARS": 1950,
         "DEVELOPMENT_SERVERS": [
-            237647756049514498,
-            815237689775357992,
-            710120382144839691
+            815237689775357992
         ]
     },
     "FILES": {
